Add files via upload
A **navbar** (short for navigation bar) is a user interface component typically found at the top or side of a webpage or app. It contains a list of links or buttons that allow users to navigate between different sections or pages within the site or application. Navbars are essential for organizing content, improving user experience, and providing easy access to key features. They can be static or dynamic, with options such as dropdown menus, search bars, and icons.